pyproject.toml
[project]
name = "instpredict"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "instructor>=1.8.3",
    "litellm>=1.71.0",
    "openai>=1.82.0",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
]
